+++ logging-hardening/Enable-Sysmon.ps1
@@ -158,7 +158,7 @@
 .NOTES
     Author  : Secur01
     Project : IronBlackBox - https://github.com/Secur01/IronBlackBox
-    Version : 1.0.1
+    Version : 1.1.0
     License : MIT
 
     Windows PowerShell 5.1. No module dependencies. Requires local
@@ -206,7 +206,7 @@ $ErrorActionPreference = 'Stop'
 Set-StrictMode -Version 1.0
 
 $script:ScriptName    = 'Enable-Sysmon'
-$script:ScriptVersion = '1.0.1'
+$script:ScriptVersion = '1.1.0'
 
 # Populated by Initialize-ToolkitRoot / Start-ManifestRun.
 $script:ManifestPath = $null